fix: bug where success screen is stuck
